cleanup and refactoring of settings menu
