+++ apps/extension/src/components/atoms/MenuButton.svelte
@@ -1,0 +1,34 @@
+<script lang="ts">
+  import { Tooltip } from "bits-ui";
+  import IconButton from "./IconButton.svelte"
+
+  type MenuButtonProps = {
+    mdiIcon: string;
+    tooltip?: string;
+    onclick?: () => void;
+  };
+
+  const { mdiIcon, tooltip, onclick }: MenuButtonProps = $props();
+</script>
+ 
+<Tooltip.Provider>
+  <Tooltip.Root delayDuration={0}>
+    <Tooltip.Trigger disabled={!tooltip}>
+      <IconButton icon={mdiIcon} onclick={onclick} />
+    </Tooltip.Trigger>
+    <Tooltip.Content sideOffset={4}>
+      <!-- putting "text-black" in order for arrow to pick it up the color and look the same as content -->
+      <Tooltip.Arrow class="text-black/50" />
+      <div
+        class={[
+          // general styles
+          'z-10 bg-black/50 text-white text-center py-1 px-2 rounded-md',
+          // animation
+          // 'opacity-0 transition-opacity', // curently not working
+        ]}
+      >
+        {tooltip}
+      </div>
+    </Tooltip.Content>
+  </Tooltip.Root>
+</Tooltip.Provider>
