+++ apps/extension/src/settings/tabs/AuthenticationTab.svelte
@@ -1,0 +1,70 @@
+<script lang="ts">
+  import AuthButton from '@/components/AuthButton.svelte'
+  import { AuthClient } from '@/oauth2/auth'
+
+  const clients = {
+    google: new AuthClient('google'),
+    spotify: new AuthClient('spotify'),
+    fitbit: new AuthClient('fitbit')
+  }
+
+  let authState = $state({
+    google: false,
+    spotify: false,
+    fitbit: false
+  })
+
+  async function retrieveAuthState() {
+    authState.google = await clients.google.isAuthenticated()
+    authState.spotify = await clients.spotify.isAuthenticated()
+    authState.fitbit = await clients.fitbit.isAuthenticated()
+  }
+</script>
+
+<h1 class="text-xl mb-3">Authentication</h1>
+{#await retrieveAuthState() then _ignore}
+  <div class="flex flex-col gap-3">
+    <p class="text-sm">
+      <strong>Google:</strong>
+      <span class="text-gray-400">{authState.google}</span>
+      {#if !authState.google}
+        <AuthButton
+          class="mt-2"
+          provider={'google'}
+          onclick={async () => {
+            const token = await clients.google.getAuthToken(true)
+            authState.google = !!token
+          }}
+        />
+      {/if}
+    </p>
+    <p class="text-sm">
+      <strong>Spotify:</strong>
+      <span class="text-gray-400">{authState.spotify}</span>
+      {#if !authState.spotify}
+        <AuthButton
+          class="mt-2"
+          provider={'spotify'}
+          onclick={async () => {
+            const token = await clients.spotify.getAuthToken(true)
+            authState.spotify = !!token
+          }}
+        />
+      {/if}
+    </p>
+    <p class="text-sm">
+      <strong>Fitbit:</strong>
+      <span class="text-gray-400">{authState.fitbit}</span>
+      {#if !authState.fitbit}
+        <AuthButton
+          class="mt-2"
+          provider={'fitbit'}
+          onclick={async () => {
+            const token = await clients.fitbit.getAuthToken(true)
+            authState.fitbit = !!token
+          }}
+        />
+      {/if}
+    </p>
+  </div>
+{/await}
