+++ apps/extension/stories/atoms/Button.stories.svelte
@@ -1,6 +1,6 @@
 <script module lang="ts">
   import { defineMeta } from '@storybook/addon-svelte-csf'
-  import Button from '@/components/Button.svelte'
+  import Button from '@/components/atoms/Button.svelte'
   import { fn } from '@storybook/test'
 
   // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
