+++ apps/extension/stories/atoms/Panel.stories.svelte
@@ -1,5 +1,5 @@
 <script module lang="ts">
-  import { defineMeta, type Args } from '@storybook/addon-svelte-csf'
+  import { defineMeta } from '@storybook/addon-svelte-csf'
   import Panel from '@/components/atoms/Panel.svelte'
   import Toggle from '@/components/atoms/Toggle.svelte'
 
@@ -16,21 +16,10 @@
 </script>
 
 <!-- More on writing stories with args: https://storybook.js.org/docs/writing-stories/args -->
-<Story name="Default">
-  {#snippet children({
-    children = 'Hello, world!',
-    ...args
-  })}
-    <Panel {...args}>{children}</Panel>
-  {/snippet}
-</Story>
+<Story name="Default">Hello World!</Story>
 
 <Story name="With Components">
-  {#snippet children(args: any)}
-    <Panel {...args}>
-      <h1 class="text-2xl font-bold">Hello, world!</h1>
-      <p class="text-base">This is a panel component.</p>
-      <Toggle parentClass="mt-2" label={'Test Label'} />
-    </Panel>
-  {/snippet}
+  <h1 class="text-2xl font-bold">Hello, world!</h1>
+  <p class="text-base">This is a panel component.</p>
+  <Toggle parentClass="mt-2" label={'Test Label'} />
 </Story>
